% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafgrs.R
\name{pafgrs_score_bivariate}
\alias{pafgrs_score_bivariate}
\title{Bivariate PA-FGRS scores}
\usage{
pafgrs_score_bivariate(prep, traits, ages, cips, h2, rg, rho_within = NULL)
}
\arguments{
\item{prep}{From [pafgrs_prepare()].}

\item{traits}{A list of two binary [trait()]s.}

\item{ages}{A list of each trait's ages, as [pafgrs_score_univariate()]
takes them.}

\item{cips}{A list of each trait's [pafgrs_cip()] table.}

\item{h2}{The two liability-scale heritabilities, each in `(0, 1]`.}

\item{rg}{Genetic correlation, in `[-1, 1]`.}

\item{rho_within}{A person's cross-trait liability correlation, in
`[-1, 1]`; `NULL` means `rg * sqrt(h2[1] * h2[2])`.}
}
\value{
A data frame with one row per proband in pedigree input-row order
  and columns `id`, `est1`, `est2`, `var1`, `var2`, `cov12`,
  `n_relatives` (people with weight > 0 on either trait), `n_obs1` and
  `n_obs2`.  With no informative relative: `est = 0`, `var = h2`,
  `cov12 = rg * sqrt(h2[1] * h2[2])`.  `attr(, "metadata")` is a list:
  `h2`, `rg`, `rho_within` (as used), and per trait `prevalence`,
  `threshold` and `controls_without_age`, then `n_probands`, `ndegree`,
  `pg_phenotype_version` and `pedigree_graph_core_rev`.
}
\description{
Scores every proband of `prep` on two genetically correlated binary
traits jointly.
}
\examples{
ped <- data.frame(id = 1:4, mother = c(NA, NA, 1, 1), father = c(NA, NA, 2, 2))
table <- pafgrs_cip(c(0, 80), c(0, 0.1))
pafgrs_score_bivariate(
  pafgrs_prepare(ped),
  traits = list(trait(c(1, 0, 0, NA)), trait(c(0, 1, NA, 0))),
  ages = list(c(50, 60, 30, NA), c(70, 40, NA, 20)),
  cips = list(table, table),
  h2 = c(0.5, 0.4), rg = 0.3
)
}
