% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgphenotype.R
\name{pgphenotype_version}
\alias{pgphenotype_version}
\title{Package version}
\usage{
pgphenotype_version()
}
\value{
A string such as `"0.1.0"`.
}
\description{
The version of the compiled Rust binding, which equals the package version.
}
\examples{
pgphenotype_version()
}
